Cast LOAD_ADDR

+++ send.c
@@ -83,7 +83,7 @@ static void send(const int sock, const char *filename) {
 				"%zu bytes/sec - "
 				"%zu%% - "
 				"%lum%lus remaining  \r",
-				offset + LOAD_ADDR,
+				offset + (size_t)LOAD_ADDR,
 				offset, totalSize,
 				bps,
 				(offset * 100) / totalSize,
